Allow sending a 255-byte payload on sendVendorNciMessage
According to the NCI specification, the payload of the

sendVendorNciMessage() API should support up to 255 bytes,

excluding the GID, OID, and LEN fields.

Bug: 456673317

Flag: EXEMPT BUGFIX

Test: atest CtsNfcTestCases

+++ libnfc-nci/src/nfa/include/nfa_dm_int.h
@@ -187,7 +187,7 @@ typedef struct {
   NFC_HDR hdr;
   tNFA_VSC_CBACK* p_cback;
   uint8_t oid;
-  uint8_t cmd_params_len;
+  uint16_t cmd_params_len;
   uint16_t pad; /* add padding to ensure the size is big enough for
                    offset=NCI_VSC_MSG_HDR_SIZE */
   uint8_t* p_cmd_params;
